requirements.txt
openai>=1.0.0
pillow>=10.0.0
openpyxl>=3.1.0
pyyaml>=6.0
